# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Servicio API para acceder a las cartas del negocio",
        "title": "The Nexus Battles II: Card API Service Reborn",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/license/mit/"
        },
        "version": "2.0"
    },
    "host": "localhost",
    "basePath": "/api",
    "paths": {
...
        "/heroes/": {
            "get": {
                "description": "este método devuelve una lista de documentos tipo Heroe",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Heroes"
                ],
                "summary": "trae una lista de documentos tipo Heroe",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Tamaño de la paginación",
                        "name": "size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Página de los documentos",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Si se activa el filtro, solo trae las cartas activas",
                        "name": "statusFilter",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de documentos tipo Heroe",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.HeroeModel"
                            }
                        }
                    },
                    "400": {
                        "description": "Parámetros de paginación inválidos",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Error interno en el servidor",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "este método recibe los parámetros de tipo Heroe y crea la carta",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Heroes"
                ],
                "summary": "crea un documento tipo Heroe",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Imagen de la carta",
                        "name": "imagen",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Icono de la carta",
                        "name": "icono",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Nombre de la carta",
                        "name": "nombre",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Clase de la carta",
                        "name": "clase",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Tipo de la carta",
                        "name": "tipo",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Poder de la carta",
                        "name": "poder",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Vida de la carta",
                        "name": "vida",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "defensa de la carta",
                        "name": "defensa",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Ataque base de la carta",
                        "name": "ataqueBase",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Ataque aleatorio de la carta",
                        "name": "ataqueRnd",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Daño de la carta",
                        "name": "daño",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Descripción de la carta",
                        "name": "descripcion",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "[Ecommerce] Precio de la carta",
                        "name": "precio",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "[Ecommerce] Descuento de la carta",
                        "name": "descuento",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "[Ecommerce] Stock de la carta",
                        "name": "stock",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Documento creado exitosamente",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Parámetros de paginación inválidos",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno en el servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "patch": {
                "description": "este método recibe los parámetros de tipo Heroe y actualiza la carta",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Heroes"
                ],
                "summary": "actualiza un documento tipo Heroe",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Id de la carta",
                        "name": "id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Imagen de la carta",
                        "name": "imagen",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Icono de la carta",
                        "name": "icono",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Nombre de la carta",
                        "name": "nombre",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "Poder de la carta",
                        "name": "poder",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "Vida de la carta",
                        "name": "vida",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "defensa de la carta",
                        "name": "defensa",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "Ataque base de la carta",
                        "name": "ataqueBase",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "Ataque aleatorio de la carta",
                        "name": "ataqueRnd",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "Daño de la carta",
                        "name": "daño",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Descripción de la carta",
                        "name": "descripcion",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "[Ecommerce] Precio de la carta",
                        "name": "precio",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "[Ecommerce] Descuento de la carta",
                        "name": "descuento",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "[Ecommerce] Stock de la carta",
                        "name": "stock",
                        "in": "formData"
                    },
                    {
                        "type": "boolean",
                        "description": "Activar o desactivar la carta",
                        "name": "estado",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Documento actualizado exitosamente",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Parámetros de paginación inválidos",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno en el servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/heroes/name/": {
            "get": {
                "description": "este método devuelve una lista de documentos tipo Heroe filtrado por nombre",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Heroes"
                ],
                "summary": "trae una lista de documentos tipo Heroe filtrado por nombre",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Tamaño de la paginación",
                        "name": "size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Página de los documentos",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "palabra clave para filtrar",
                        "name": "keyword",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de documentos tipo Heroe",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.HeroeModel"
                            }
                        }
                    },
                    "400": {
                        "description": "Parámetros de paginación inválidos",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Error interno en el servidor",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.HeroeModel": {
            "type": "object",
            "required": [
                "ataqueBase",
                "ataqueRnd",
                "clase",
                "coleccion",
                "daño",
                "defensa",
                "descripcion",
                "descuento",
                "estado",
                "icono",
                "imagen",
                "nombre",
                "poder",
                "precio",
                "stock",
                "tipo",
                "vida"
            ],
            "properties": {
                "ataqueBase": {
                    "type": "integer"
                },
                "ataqueRnd": {
                    "type": "integer"
                },
                "clase": {
                    "type": "string",
                    "enum": [
                        "Guerrero",
                        "Mago",
                        "Pícaro"
                    ]
                },
                "coleccion": {
                    "type": "string",
                    "enum": [
                        "Heroes",
                        "Armas",
                        "Armaduras",
                        "Items",
                        "Épicas"
                    ]
                },
                "daño": {
                    "type": "integer"
                },
                "defensa": {
                    "type": "integer"
                },
                "descripcion": {
                    "type": "string"
                },
                "descuento": {
                    "type": "integer",
                    "minimum": 0
                },
                "estado": {
                    "type": "boolean"
                },
                "icono": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "imagen": {
                    "type": "string"
                },
                "nombre": {
                    "type": "string"
                },
                "poder": {
                    "type": "integer"
                },
                "precio": {
                    "type": "integer",
                    "minimum": 0
                },
                "stock": {
                    "type": "integer",
                    "minimum": 0
                },
                "tipo": {
                    "type": "string",
                    "enum": [
                        "Tanque",
                        "Armas",
                        "Fuego",
                        "Hielo",
                        "Veneno",
                        "Machete"
                    ]
                },
                "vida": {
                    "type": "integer"
                }
            }
        }
    },
...
}